Implementing Atomic Design v.1.2: All concept is implemented
