--- src/components/atoms/Button/button.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { Button } from "react-bootstrap"
-
-const Button = (props) => {
-	return (
-		<Button>{props.buttonText}</Button>
-	);
-};
-
-export default Button;
