+++ src/components/atoms/Rating/style.scss
@@ -1,0 +1,7 @@
+.rating {
+    display: flex;
+    justify-content: space-between;
+    .rt_score{
+        margin-left: 3px;
+    }
+}
