--- src/components/atoms/Title/title.jsx
@@ -1,10 +1,0 @@
-import React from "react"
-import { Card } from "react-bootstrap";
-
-const Title = (props) => {
-    return(
-        <Card.Title>{props.title}</Card.Title>
-    )
-}
-
-export default Title;
