+++ src/components/molecules/Sublabel/style.scss
@@ -1,0 +1,7 @@
+.sublabel {
+    display: flex;
+    .year {
+        color: grey;
+        margin-left: 7px;
+    }
+}
