+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import Main from './components/Main';
+import Main from './pages/Main';
 
 ReactDOM.render(
     <Main />,
