+++ src/pages/People.js
