+++ src/styles.scss
@@ -157,6 +157,10 @@ $fontFam: 'Montserrat', sans-serif;
 
 }
 
+.error {
+    color: $firstTextColor;
+}
+
 .overflow-table {
     overflow-x:auto
 }
